Move various config parsing functions from Core and Runner into ConfigParser.
