+++ lighthouse-core/runner.js
@@ -24,27 +24,8 @@ const fs = require('fs');
 const path = require('path');
 
 class Runner {
-
-  static getGatherersNeededByAudits(audits) {
-    // It's possible we didn't get given any audits (but existing audit results), in which case
-    // there is no need to do any work here.
-    if (!audits) {
-      return new Set();
-    }
-
-    audits = Core.expandAudits(audits);
-
-    return audits.reduce((list, audit) => {
-      audit.meta.requiredArtifacts.forEach(artifact => list.add(artifact));
-      return list;
-    }, new Set());
-  }
-
   static run(driver, opts) {
     const config = opts.config;
-
-    // Filter out any audits by the whitelist.
-    config.audits = Core.filterAudits(config.audits, opts.flags.auditWhitelist);
 
     // Check that there are passes & audits...
     const validPassesAndAudits = (
@@ -63,32 +44,8 @@ class Runner {
     // to check that there are artifacts specified in the config, and throw if not.
     if (validPassesAndAudits || validArtifactsAndAudits) {
       if (validPassesAndAudits) {
-        const requiredGatherers = this.getGatherersNeededByAudits(config.audits);
-
-        // Make sure we only have the gatherers that are needed by the audits
-        // that have been listed in the config.
-        const filteredPasses = config.passes.map(pass => {
-          pass.gatherers = pass.gatherers.filter(gatherer => {
-            if (typeof gatherer !== 'string') {
-              return requiredGatherers.has(gatherer.name);
-            }
-
-            try {
-              const GathererClass = Driver.getGathererClass(gatherer);
-              return requiredGatherers.has(GathererClass.name);
-            } catch (requireError) {
-              throw new Error(`Unable to locate gatherer: ${gatherer}`);
-            }
-          });
-
-          return pass;
-        })
-
-        // Now remove any passes which no longer have gatherers.
-        .filter(p => p.gatherers.length > 0);
-
         // Finally set up the driver to gather.
-        run = run.then(_ => Driver.run(filteredPasses, Object.assign({}, opts, {driver})));
+        run = run.then(_ => Driver.run(config.passes, Object.assign({}, opts, {driver})));
       } else if (validArtifactsAndAudits) {
         run = run.then(_ => config.artifacts);
       }
